backend/requirements.txt
# =============================================================================
# KARI.Самозанятые — Зависимости Python
# Установка: pip install -r requirements.txt
# =============================================================================

# --- Веб-фреймворк ---
fastapi==0.115.5             # Основной фреймворк API
uvicorn[standard]==0.32.1    # Сервер для запуска FastAPI

# --- База данных ---
sqlalchemy[asyncio]==2.0.36  # ORM для работы с PostgreSQL
asyncpg==0.30.0              # Асинхронный драйвер PostgreSQL
alembic==1.14.0              # Миграции базы данных

# --- Настройки ---
pydantic-settings==2.6.1     # Чтение настроек из .env файла
python-dotenv==1.0.1         # Загрузка .env файла

# --- Авторизация ---
python-jose[cryptography]==3.3.0  # JWT токены
passlib[bcrypt]==1.7.4            # Хэширование (не используем пароли, но нужно для токенов)

# --- HTTP-запросы (к ФНС, Совкомбанку) ---
httpx==0.28.1                # Асинхронный HTTP-клиент

# --- Redis и очереди ---
redis==5.2.1                 # Подключение к Redis
celery==5.4.0                # Очереди задач (выплаты, проверка чеков)

# --- Хранилище файлов ---
minio==7.2.12                # Клиент MinIO (фотоотчёты)

# --- Push-уведомления ---
firebase-admin==6.6.0        # Firebase Cloud Messaging

# --- Работа с файлами ---
python-multipart==0.0.20     # Загрузка файлов (фотоотчёты)
openpyxl==3.1.5              # Чтение/запись Excel (реестры выплат)
Pillow==11.0.0               # Обработка изображений (проверка размера фото)

# --- Дата и время ---
python-dateutil==2.9.0       # Удобная работа с датами

# --- Генерация PDF (договоры ГПХ и акты выполненных работ) ---
reportlab==4.2.5             # Генерация PDF-документов (чистый Python)

# --- Утилиты ---
phonenumbers==8.13.50        # Валидация российских номеров телефонов
